requirementsold.txt
pandas==2.0.3
seaborn==0.13.1
streamlit==1.29.0
numpy==1.24.3
matplotlib.pyplot==3.7.2